AB#44796

+++ Models/BenchmarkChartData.cs
@@ -51,6 +51,6 @@ namespace SFB.Web.ApplicationCore.Models
         public string ProgressScoreType { get; set; }
         
         [JsonProperty(PropertyName = "topsen")]
-        public List<KeyValuePair<string, int>> TopSEN { get; set; }
+        public List<KeyValuePair<string, decimal>> TopSEN { get; set; }
     }
 }
